Stage 3

+++ Recipes/task/src/recipes/models/DTOs.java
@@ -1,7 +1,7 @@
 package recipes.models;
 
 public class DTOs {
-    public record RecipeIdDTO(int id) {
+    public record RecipeIdDTO(Long id) {
     }
 
 }
